Add functions for keyboard as input
Extra credit: add backspace button and also
keyboard now takes in as input alternative.

- used something i haven't the faintest idea
what it is for: Event: preventDefault() method

I added this to the window.keydown listener.

I copied and pasted the functions from the calculator
divs to the window keydown listener.

Also fixed the fact that when you backspace after
pressing = or any other operators ; the calculator
will empty and break.